backend/app/core/security.py
"""
Security utilities for authentication and authorization.

Provides password hashing and JWT token utilities.
Note: Primary authentication is handled by Supabase Auth.
These utilities are for additional backend security if needed.
"""

import os
from datetime import datetime, timedelta
from typing import Any

from jose import JWTError, jwt
from passlib.context import CryptContext

# Password hashing context
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# JWT configuration
SECRET_KEY = os.getenv("JWT_SECRET_KEY", "your-secret-key-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "60"))


def hash_password(password: str) -> str:
    """
    Hash a password using bcrypt.

    Args:
        password: Plain text password

    Returns:
        Hashed password string
    """
    return pwd_context.hash(password)


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    Verify a password against a hash.

    Args:
        plain_password: Plain text password
        hashed_password: Hashed password to verify against

    Returns:
        True if password matches hash
    """
    return pwd_context.verify(plain_password, hashed_password)


def create_access_token(data: dict[str, Any], expires_delta: timedelta | None = None) -> str:
    """
    Create a JWT access token.

    Args:
        data: Data to encode in the token
        expires_delta: Token expiration time (default: ACCESS_TOKEN_EXPIRE_MINUTES)

    Returns:
        Encoded JWT token string
    """
    to_encode = data.copy()

    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

    return encoded_jwt


def decode_access_token(token: str) -> dict[str, Any] | None:
    """
    Decode and verify a JWT access token.

    Args:
        token: JWT token string

    Returns:
        Decoded token payload or None if invalid
    """
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None


def verify_supabase_jwt(token: str) -> dict[str, Any] | None:
    """
    Verify a Supabase JWT token.

    Supabase uses JWT tokens for authentication. This function verifies
    the token signature and extracts the user information.

    Args:
        token: Supabase JWT token

    Returns:
        Decoded token payload or None if invalid

    Note:
        In production, you should use Supabase's JWT secret from environment variables.
        The secret is available in your Supabase project settings under API > JWT Secret.
    """
    supabase_jwt_secret = os.getenv("SUPABASE_JWT_SECRET", SECRET_KEY)

    try:
        payload = jwt.decode(token, supabase_jwt_secret, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
