backend/pyproject.toml
[project]
name = "efir-budget-backend"
version = "0.1.0"
description = "FastAPI service for EFIR Budget Planning"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "EFIR", email = "tech@efir.local"}]
dependencies = [
  "fastapi==0.123.0",
  "uvicorn[standard]==0.34.0",
  "pydantic==2.12.0",
  "sqlalchemy==2.0.36",
  "asyncpg==0.30.0",
  "python-dotenv==1.0.1",
  "passlib[bcrypt]==1.7.4",
  "python-jose[cryptography]==3.3.0"
]

[project.optional-dependencies]
dev = [
  "ruff==0.8.2",
  "mypy==1.14.1",
  "pytest==8.3.3",
  "httpx==0.27.2",
  "pytest-asyncio==0.24.0"
]

[tool.ruff]
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "RUF", "ASYNC"]
ignore = [
    "RUF001", "RUF002", "RUF003",  # Allow × (multiplication sign) in strings/comments/docstrings
    "RUF012",  # Pydantic Config class uses mutable defaults intentionally
    "B008",  # Depends() in defaults is standard FastAPI pattern
    "B904",  # Don't chain HTTPExceptions to avoid exposing internal errors
]

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"]  # Imports after sys.path manipulation is standard Alembic pattern
"alembic/versions/*" = ["E501"]  # Allow long lines in migration files (auto-generated)
"tests/test_models.py" = ["F406"]  # Allow star import in test function
"app/services/*.py" = ["E501"]  # Allow longer lines for descriptive error messages
"app/api/**/*.py" = ["E501"]  # Allow longer lines in API schemas
"app/schemas/*.py" = ["E501"]  # Allow longer lines for Field descriptions
"app/engine/**/*.py" = ["E501"]  # Allow longer lines in engine models
"app/core/*.py" = ["E501"]  # Allow longer lines in core modules
"app/models/*.py" = ["E501"]  # Allow longer lines for SQL comments

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = false  # Too strict for evolving codebase
strict_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["app.models.base"]
# SQLAlchemy Base is dynamically created, ignore typing issues
ignore_errors = true

[[tool.mypy.overrides]]
module = ["app.services.*"]
# Service modules under development, will be fixed in dedicated pass
ignore_errors = true

[[tool.mypy.overrides]]
module = ["app.api.v1.*"]
# API routes may reference not-yet-exported functions
disable_error_code = ["attr-defined"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["backend/tests"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
